requirements.txt
torch>=2.0.0
openai-whisper>=20231117
transformers>=4.37.0
pillow>=10.0.0
accelerate>=0.26.0 